LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/messagechannel.any.worker-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: window

Harness Error (FAIL), message = ReferenceError: Can't find variable: window
